refactor: application 패키지를 제거하도록 변경
