+++ src/main/java/roomescape/controller/ReservationController.java
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.service.ReservationService;
 import roomescape.controller.request.CreateReservationWebRequest;
 import roomescape.controller.response.ReservationWebResponse;
 import roomescape.domain.Reservation;
+import roomescape.service.ReservationService;
 
 @RestController
 @RequestMapping("/reservations")
