+++ src/main/java/roomescape/controller/request/CreateReservationWebRequest.java
@@ -3,7 +3,7 @@ package roomescape.controller.request;
 import java.time.LocalDate;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import roomescape.application.request.CreateReservationRequest;
+import roomescape.service.request.CreateReservationRequest;
 
 public record CreateReservationWebRequest(
